Move user plugin to plugin project
