+++ Flow.Launcher.Core/ExternalPlugins/PluginsManifest.cs
@@ -1,4 +1,5 @@
 ﻿using Flow.Launcher.Infrastructure.Logger;
+using Flow.Launcher.Plugin;
 using System;
 using System.Collections.Generic;
 using System.Threading;
