+++ Flow.Launcher/ViewModel/PluginStoreItemViewModel.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Linq;
 using CommunityToolkit.Mvvm.Input;
-using Flow.Launcher.Core.ExternalPlugins;
 using Flow.Launcher.Core.Plugin;
 using Flow.Launcher.Plugin;
-using SemanticVersioning;
 using Version = SemanticVersioning.Version;
 
 namespace Flow.Launcher.ViewModel
